Se cambia la conexión MySQL por PostgreSQL. Creación de modelos y controladores.

+++ src/migrations/20191230235900-create-users.js
@@ -29,7 +29,7 @@ module.exports = {
         defaultValue: '',
       },
       is_enabled: {
-        type: Sequelize.TINYINT(1),
+        type: Sequelize.INTEGER,
         allowNull: false,
         defaultValue: 1,
       },
